sellersco-worker/requirements-evaluation.txt
azure-ai-evaluation>=0.3.0
azure-identity>=1.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
